Allow multiple local KMLs to be loaded
Loading more than one local KML caused an exception to be thrown and
broke the KML import tool for both local and distant KMLs.

To reproduce: import a local KML from your hard drive and click on the
'Load KML' button twice or try to load another KML.

This seems to come from the fact that before this commit, once a local
KML was loaded, the fileReader reference was always truethy which means
in $scope.cancel we were trying to abort the upload of an uploaded
file. This caused the exception.