Change size of printed texts
We used to rely on the size of OpenLayers and not to take the scale of
the text (small, medium, big) into account. We now use a smaller font
size (10px) which looks better on the PDF and multiply it by the scale.

We also reduced the size of the halo.

Both options can be changed with print.base_text_halo_size and
print.base_text_size.